+++ test/e2e/test-app/fixtures/sentry-basic.js
@@ -1,4 +1,4 @@
-const { init, captureMessage } = require('../../../../');
+const { init } = require('../../../../');
 
 init({
   dsn: process.env.DSN,