+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased
+
+**Breaking Change**: This version uses the [envelope endpoint](https://develop.sentry.dev/sdk/envelopes/). If you are
+using an on-premise installation it requires Sentry version `>= v20.6.0` to work. If you are using
+[sentry.io](https://sentry.io) nothing will change and no action is needed.
+
+- ref: Decrease bundle size by removing dependencies (#252)
+- ref: Use envelope endpoint (#252)
+- feat: Export NetTransport (#252)
+- feat: Export `flush` & `close` (#252)
+
 ## v1.5.2
 
 - fix: Cirular refs (#253)